Move check method to uvbase
